mch day chart
Former-commit-id: 9ec498821b07e4f61e4dd1082bd0d1c73dfeaa8b

+++ core/domain/interface/merchant/merchant.go
@@ -217,7 +217,7 @@ type (
 		// 线下订单金额
 		OfflineAmount float32 `db:"offline_amount"`
 		// 日期
-		Date int `db:"date"`
+		Date int64 `db:"date"`
 		// 日期字符串
 		DateStr string `db:"date_str"`
 		// 更新时间
